renaming files structure in area proj
