+++ C#.NET Apps/YouTubeProjects/YTP.Main/App_Start/BundleConfig.cs
@@ -13,17 +13,17 @@ namespace YTP.Main {
                       "~/GlobalAssets/Stylings/site.css"));
 
             bundles.Add(new ScriptBundle("~/bundles/jquery").Include(
-                        "~/GlobalAssets/ScriptFiles/jquery-{version}.js"));
+                      "~/GlobalAssets/ScriptFiles/jquery-{version}.js"));
 
             bundles.Add(new ScriptBundle("~/bundles/jqueryval").Include(
-                        "~/GlobalAssets/ScriptFiles/jquery.validate*"));
+                      "~/GlobalAssets/ScriptFiles/jquery.validate*"));
 
             bundles.Add(new ScriptBundle("~/bundles/jqueryUI").Include(
-                        "~/GlobalAssets/ScriptFiles/jquery-ui-1.13.2.min.js"));
+                      "~/GlobalAssets/ScriptFiles/jquery-ui-1.13.2.min.js"));
 
             bundles.Add(new ScriptBundle("~/bundles/DataTables").Include(
-                        "~/GlobalAssets/ScriptFiles/jquery.dataTables.min.js",
-                        "~/GlobalAssets/ScriptFiles/dataTables.bootstrap5.min.js"));
+                      "~/GlobalAssets/ScriptFiles/jquery.dataTables.min.js",
+                      "~/GlobalAssets/ScriptFiles/dataTables.bootstrap5.min.js"));
 
             bundles.Add(new Bundle("~/bundles/bootstrap").Include(
                       "~/GlobalAssets/ScriptFiles/bootstrap.bundle.js"));
