+++ C#.NET Apps/YouTubeProjects/YTP.Main/Areas/HumanResources/LocalAssets/haitiemployee.js
@@ -1,6 +1,6 @@
 ﻿//Load Data in Table when documents is ready
 
-var mainUrl = '/HaitiEmployee/HaitiEmployee';
+var mainUrl = '/HumanResources/HaitiEmployee';
 $(document).ready(function () {
     loadData();
 });
